Create Tasks container and rename ApiCompatTask folder
