+++ Tasks/ApiCompat/tests/3.20.1/Microsoft.Bot.Builder.Autofac.dll
