+++ Tasks/ApiCompat/tests/4.3.2/Microsoft.Bot.Builder.dll
